modify: move components file
