+++ browser/components/NoteItem.js
