+++ browser/components/NoteItem.styl
