+++ browser/components/NoteItemSimple.js
