+++ browser/components/NoteItemSimple.styl
